자잘한 기능들 추가 및 수정

+++ static/js/delete.js
@@ -1,6 +1,8 @@
 $(".btn-delete-post").click(() => {
-    const post_idx = $("input[name='post-idx']").val();
-    deleteData(`/post/${post_idx}`)
+    if(confirm("해당 게시글을 삭제 하겠습니까?")){
+        const post_idx = $("input[name='post-idx']").val();
+        deleteData(`/post/${post_idx}`);
+    }
 })
 
 function deleteData(url){
